Allow clauses config to accept symbols like :set, :not_set
We're validating them against our known Clauses names, so we don't have mispellings.

By allowing to refer to clauses as symbols directly, we remove the need for all the const reassignments in conditions.

Also adds `without_presence_clauses` because that's often used.

Apps will have to swap to symbols immediately on upgrade, we could add a backwardscompatibility layer though.